Create UrlShortener.py
In this code, we define a Flask application and create routes to handle different requests. When a user enters a long URL in the form on the homepage, we generate a short code and store the mapping between the short code and long URL in a dictionary. We then display the shortened URL to the user. When a user enters the shortened URL in their browser, we use the short code to look up the corresponding long URL and redirect the user to that page.